Fix semantic clone fixture current projections
