+++ bitloops/src/host/devql/tests/cucumber_bdd.rs
@@ -1,6 +1,7 @@
 use super::cucumber_steps;
 use super::cucumber_world::DevqlBddWorld;
-use cucumber::{World as _, writer::Stats as _};
+use cucumber::{World as _, event::ScenarioFinished, writer::Stats as _};
+use std::sync::{Arc, Mutex};
 
 #[tokio::test]
 async fn devql_bdd_features_pass() {
@@ -8,6 +9,8 @@ async fn devql_bdd_features_pass() {
         "{}/src/host/devql/tests/features",
         env!("CARGO_MANIFEST_DIR")
     );
+    let failed_scenarios = Arc::new(Mutex::new(Vec::new()));
+    let failed_scenarios_for_after = Arc::clone(&failed_scenarios);
 
     let result = DevqlBddWorld::cucumber()
         .steps(cucumber_steps::collection())
@@ -21,14 +24,37 @@ async fn devql_bdd_features_pass() {
                 world.scenario_id = scenario.name.split_whitespace().next().map(str::to_string);
             })
         })
+        .after(move |_, _, scenario, event, _| {
+            let failed_scenarios = Arc::clone(&failed_scenarios_for_after);
+            Box::pin(async move {
+                if matches!(
+                    event,
+                    ScenarioFinished::BeforeHookFailed(_) | ScenarioFinished::StepFailed(_, _, _)
+                ) {
+                    failed_scenarios
+                        .lock()
+                        .expect("failed cucumber scenario list lock")
+                        .push(scenario.name.clone());
+                }
+            })
+        })
         .with_default_cli()
         .fail_on_skipped()
         .run(feature_dir)
         .await;
+    let failed_scenarios = failed_scenarios
+        .lock()
+        .expect("failed cucumber scenario list lock")
+        .join(", ");
 
     assert!(
         !result.execution_has_failed(),
-        "cucumber suite reported failures"
+        "cucumber suite reported failures{}",
+        if failed_scenarios.is_empty() {
+            String::new()
+        } else {
+            format!(": {failed_scenarios}")
+        }
     );
     assert_eq!(result.skipped_steps(), 0, "cucumber suite skipped steps");
     assert_eq!(
